分页代码修复

+++ src/main/java/com/cudrania/hibernate/Page.java
@@ -90,7 +90,7 @@ public interface Page<T> {
      *
      * @return
      */
-    long getTotalPages();
+    int getTotalPages();
 
     /**
      * 是否有下一页
